Fix typo in throttled request info log
These log lines previously began with "Throttling", as if we were
informing the user that future requests would be throttled. However,
these info logs are a report of the request latency due to the
existing request throttling, and so the verb phrase in these log lines
("Throtting request") was replaced with the noun prase "Throttled
request".